--- app/src/main/java/com/leoapps/findout/add/navigation/model/AddNavCommand.kt
@@ -1,8 +1,0 @@
-package com.leoapps.findout.add.navigation.model
-
-sealed interface AddNavCommand {
-    object GoBack : AddNavCommand
-    data class OpenAddQuestionScreen(
-        val isEdit: Boolean
-    ) : AddNavCommand
-}
