--- app/src/main/java/com/leoapps/findout/add_question/navigation/model/AddQuestionNavCommand.kt
@@ -1,5 +1,0 @@
-package com.leoapps.findout.add_question.navigation.model
-
-sealed interface AddQuestionNavCommand {
-    object GoBack : AddQuestionNavCommand
-}
