+++ app/src/main/java/com/leoapps/findout/creation/form/navigation/model/FormCreationNavCommand.kt
@@ -1,0 +1,8 @@
+package com.leoapps.findout.creation.form.navigation.model
+
+sealed interface FormCreationNavCommand {
+    object GoBack : FormCreationNavCommand
+    data class OpenAddQuestionScreen(
+        val isEdit: Boolean
+    ) : FormCreationNavCommand
+}
