+++ app/src/main/java/com/leoapps/findout/creation/question/navigation/model/QuestionCreationNavCommand.kt
@@ -1,0 +1,5 @@
+package com.leoapps.findout.creation.question.navigation.model
+
+sealed interface QuestionCreationNavCommand {
+    object GoBack : QuestionCreationNavCommand
+}
