+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,11 @@
 <resources>
     <string name="app_name">FindOut</string>
 
+    <string name="form_creation_title_quiz">Create Quiz</string>
+    <string name="form_creation_title_survey">Create Survey</string>
+
+    <string name="answer_creation_input_placeholder">Enter the answer</string>
+
     <string name="question_type_single">Single Answer</string>
     <string name="question_type_single_choice">Single Choice</string>
     <string name="question_type_multiple">Multiple Answer</string>
